Client functionality
- Client now works correctly
