+++ client/src/controllers/index.js
@@ -1,0 +1,6 @@
+
+var app = require("angular").module('TradeNet');
+
+app.controller('loginController',   require('./login-controller.js'));
+app.controller('logoutController',   require('./logout-controller.js'));
+app.controller('registerController',   require('./register-controller.js'));
