+++ client/src/controllers/login-controller.js
@@ -1,0 +1,32 @@
+module.exports = loginController;
+
+/* @ngInject */
+function loginController($scope, $location, AuthService) {
+
+    console.log("Login controller up!");
+
+    $scope.login = function() {
+
+        // initial values
+        $scope.error = false;
+        $scope.disabled = true;
+
+        // call login from service
+        AuthService.login($scope.loginForm.username, $scope.loginForm.password)
+            // handle success
+            .then(function() {
+                $location.path('/');
+                $scope.disabled = false;
+                $scope.loginForm = {};
+            })
+            // handle error
+            .catch(function() {
+                $scope.error = true;
+                $scope.errorMessage = "Invalid username and/or password";
+                $scope.disabled = false;
+                $scope.loginForm = {};
+            });
+
+    };
+
+}
