+++ client/src/directives/symbol-info.js
