+++ client/src/index.html
@@ -1,4 +1,4 @@
-<html ng-app="TradeNet">
+<html lang="en">
 
 <head>
     <title>TradeNet</title>
@@ -14,14 +14,12 @@
         <div ng-view></div>
     </div>
 
+    <p>this is index.html</p>
+
+    <tn-symbol-info md-whiteframe="3" style="padding:10px;"></tn-symbol-info>
+
     <!-- Scripts -->
     <script src="//code.jquery.com/jquery-2.2.1.min.js"></script>
-    <script src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/angular.js/1.4.9/angular.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/angular.js/1.4.9/angular-route.min.js"></script>
-    <script src="./main.js"></script>
-    <script src="./services.js"></script>
-    <script src="./controllers.js"></script>
     <!-- t.c. Should this be server/app.js instead? -->
     <script src="app.js"></script>
 </body>
