+++ client/src/partial/home.html
@@ -2,4 +2,6 @@
     <a ng-click='logout()' class="btn btn-default">Log Out</a>
 </div>
 
+<p>this is home.html</p>
+
 <tn-symbol-info md-whiteframe="3" style="padding:10px;"></tn-symbol-info>
