+++ client/src/services/index.js
@@ -2,3 +2,4 @@
 var app = require("angular").module('TradeNet');
 
 app.service('Server',   require('./Server.js'));
+app.factory('AuthService',   require('./login-services.js'));
