+++ package.json
@@ -1,53 +1,55 @@
 {
-    "name": "se-arch-final",
-    "version": "1.0.0",
-    "description": "Byron's end of semester group assignment",
-    "main": "./server/main.js",
-    "scripts": {
-        "start": "node ./server/main.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "keywords": [
-        "School",
-        "Software",
-        "Architecture"
-    ],
-    "author": "Eli Davis",
-    "license": "MIT",
-    "dependencies": {
-        "angular": "^1.5.8",
-        "angular-animate": "^1.5.8",
-        "angular-aria": "^1.5.8",
-        "angular-material": "^1.1.1",
-        "bcryptjs": "*",
-        "body-parser": "*",
-        "browserify": "^13.0.1",
-        "browserify-ngannotate": "^2.0.0",
-        "connect-flash": "*",
-        "cookie-parser": "^1.4.1",
-        "express": "^4.14.0",
-        "express-handlebars": "*",
-        "express-messages": "*",
-        "express-session": "*",
-        "express-validator": "*",
-        "fs-jetpack": "^0.9.2",
-        "gulp": "^3.9.1",
-        "gulp-concat": "^2.6.0",
-        "gulp-minify": "0.0.12",
-        "gulp-sass": "^2.3.2",
-        "gulp-streamify": "^1.0.2",
-        "gulp-uglify": "^1.5.3",
-        "mongodb": "^2.2.11",
-        "mongoose": "*",
-        "passport": "*",
-        "passport-http": "*",
-        "passport-local": "*",
-        "passport-local-mongoose": "^4.0.0",
-        "rx": "^4.1.0",
-        "rx-angular": "^1.1.3",
-        "socket.io": "^1.4.8",
-        "vinyl-source-stream": "^1.1.0",
-        "debug": "^2.2.0",
-        "morgan": "^1.7.0"
-    }
+  "name": "se-arch-final",
+  "version": "1.0.0",
+  "description": "Byron's end of semester group assignment",
+  "main": "./server/app.js",
+  "scripts": {
+    "start": "node ./server/app.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [
+    "School",
+    "Software",
+    "Architecture"
+  ],
+  "author": "Eli Davis",
+  "license": "MIT",
+  "dependencies": {
+    "angular": "^1.5.8",
+    "angular-animate": "^1.5.8",
+    "angular-aria": "^1.5.8",
+    "angular-material": "^1.1.1",
+    "angular-route": "^1.5.8",
+    "bcrypt-nodejs": "0.0.3",
+    "bcryptjs": "*",
+    "body-parser": "*",
+    "browserify": "^13.0.1",
+    "browserify-ngannotate": "^2.0.0",
+    "connect-flash": "*",
+    "cookie-parser": "^1.4.1",
+    "debug": "^2.2.0",
+    "express": "^4.14.0",
+    "express-handlebars": "*",
+    "express-messages": "*",
+    "express-session": "*",
+    "express-validator": "*",
+    "fs-jetpack": "^0.9.2",
+    "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.0",
+    "gulp-minify": "0.0.12",
+    "gulp-sass": "^2.3.2",
+    "gulp-streamify": "^1.0.2",
+    "gulp-uglify": "^1.5.3",
+    "mongodb": "^2.2.11",
+    "mongoose": "^4.4.5",
+    "morgan": "^1.7.0",
+    "passport": "*",
+    "passport-http": "*",
+    "passport-local": "*",
+    "passport-local-mongoose": "^4.0.0",
+    "rx": "^4.1.0",
+    "rx-angular": "^1.1.3",
+    "socket.io": "^1.4.8",
+    "vinyl-source-stream": "^1.1.0"
+  }
 }
